Changed estimatedValue to BigDecimal, commented out CreateDealRequestDto assignedManagerId since its in param, added securityConfig to allow testing with MongoDB Compass

+++ backend/investment-deal-room-backend/src/main/java/com/skillstorm/investment_deal_room_backend/dtos/dealDtos/request/UpdateDealRequestDto.java
@@ -1,6 +1,8 @@
 package com.skillstorm.investment_deal_room_backend.dtos.dealDtos.request;
 
 
+import java.math.BigDecimal;
+
 import com.skillstorm.investment_deal_room_backend.enums.Currency;
 import com.skillstorm.investment_deal_room_backend.enums.DealType;
 
@@ -8,7 +10,7 @@ public record UpdateDealRequestDto(
     String dealName,
     DealType dealType,
     String targetCompany,
-    Long estimatedValue,
+    BigDecimal estimatedValue,
     Currency currency
 ) {}
